rename class ApiPieDoc::Attribute
