+++ lib/simple_params.rb
@@ -6,3 +6,4 @@ require 'simple_params/validations'
 require 'simple_params/params'
 require 'simple_params/simple_params_error'
 require 'simple_params/api_pie_doc'
+require 'simple_params/api_pie_doc/attribute'
